+++ scripts/run_examples.sh
@@ -21,6 +21,7 @@ track_sequence_with_extractor_dummy
 track_sequence_with_gnss_prior
 track_sequence_with_two_view_match_vo_prior
 track_sequence_with_visual_odometry_prior
+two_view_vo_compare
 visual_odometry_prior_dummy
 "